Add GetRemoved User Service

+++ Src/KalaMarket/KalaMarket.Shared/Pagination.cs
@@ -2,7 +2,7 @@
 
 public static class Pagination
 {
-    public static IEnumerable<TSource> ToPaged<TSource>(this IEnumerable<TSource> source, int page, int pageSize,
+    public static IEnumerable<TSource> ToPaged<TSource>(this IEnumerable<TSource> source, int page, byte pageSize,
         out int rowsCounts)
     {
         rowsCounts = source.Count();
